+++ demisto_sdk/commands/common/tests/base_validator_test.py
@@ -1,7 +1,6 @@
 from os.path import join
 
-from demisto_sdk.commands.common.constants import (PACK_METADATA_CERTIFICATION,
-                                                   PACK_METADATA_SUPPORT)
+from demisto_sdk.commands.common.constants import PACK_METADATA_SUPPORT
 from demisto_sdk.commands.common.errors import (FOUND_FILES_AND_ERRORS,
                                                 FOUND_FILES_AND_IGNORED_ERRORS,
                                                 PRESET_ERROR_TO_CHECK,
@@ -179,8 +178,7 @@ def test_check_support_status_xsoar_file(repo, mocker):
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
     meta_json = {
-        PACK_METADATA_SUPPORT: "xsoar",
-        PACK_METADATA_CERTIFICATION: "certified"
+        PACK_METADATA_SUPPORT: "xsoar"
     }
     mocker.patch.object(BaseValidator, 'get_metadata_file_content', return_value=meta_json)
     pack.pack_metadata.write_json(meta_json)
@@ -191,22 +189,21 @@ def test_check_support_status_xsoar_file(repo, mocker):
         assert 'integration.yml' not in base_validator.ignored_errors
 
 
-def test_check_support_status_non_certified_partner_file(repo, mocker):
+def test_check_support_status_partner_file(repo, mocker):
     """
     Given
-    - An non-certified partner supported integration yml.
+    - An partner supported integration yml.
 
     When
     - Running check_support_status method.
 
     Then
-    - Ensure the resulting ignored errors list includes the non-certified-partner ignore-list.
+    - Ensure the resulting ignored errors list includes the partner ignore-list.
     """
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
     meta_json = {
-        PACK_METADATA_SUPPORT: "partner",
-        PACK_METADATA_CERTIFICATION: "not certified"
+        PACK_METADATA_SUPPORT: "partner"
     }
     mocker.patch.object(BaseValidator, 'get_metadata_file_content', return_value=meta_json)
     pack.pack_metadata.write_json(meta_json)
@@ -214,33 +211,7 @@ def test_check_support_status_non_certified_partner_file(repo, mocker):
         base_validator = BaseValidator(ignored_errors={})
         base_validator.update_checked_flags_by_support_level(integration.yml.rel_path)
 
-        assert base_validator.ignored_errors['integration.yml'] == PRESET_ERROR_TO_IGNORE['non-certified-partner']
-
-
-def test_check_support_status_certified_partner_file(repo, mocker):
-    """
-    Given
-    - An certified partner supported integration yml.
-
-    When
-    - Running check_support_status method.
-
-    Then
-    - Ensure the resulting ignored errors list includes the community ignore-list.
-    """
-    pack = repo.create_pack('pack')
-    integration = pack.create_integration('integration')
-    meta_json = {
-        PACK_METADATA_SUPPORT: "partner",
-        PACK_METADATA_CERTIFICATION: "certified"
-    }
-    mocker.patch.object(BaseValidator, 'get_metadata_file_content', return_value=meta_json)
-    pack.pack_metadata.write_json(meta_json)
-    with ChangeCWD(repo.path):
-        base_validator = BaseValidator(ignored_errors={})
-        base_validator.update_checked_flags_by_support_level(integration.yml.rel_path)
-
-        assert 'integration.yml' not in base_validator.ignored_errors
+        assert base_validator.ignored_errors['integration.yml'] == PRESET_ERROR_TO_IGNORE['partner']
 
 
 def test_check_support_status_community_file(repo, mocker):
@@ -257,8 +228,7 @@ def test_check_support_status_community_file(repo, mocker):
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
     meta_json = {
-        PACK_METADATA_SUPPORT: "community",
-        PACK_METADATA_CERTIFICATION: "not certified"
+        PACK_METADATA_SUPPORT: "community"
     }
     mocker.patch.object(BaseValidator, 'get_metadata_file_content', return_value=meta_json)
     pack.pack_metadata.write_json(meta_json)